Seperated the '/movies/movieId' route into it's own subdirectory
