+++ src/movies/movieId/movieId.service.js
@@ -1,0 +1,11 @@
+const db = require("../../db/connection");
+const tableName = "movies";
+
+/**************************** CRUDL Knex Queries ****************************/
+function read(movie_id) {
+  return db(tableName).select("*").where({ movie_id }).first();
+}
+
+module.exports = {
+  read,
+};
